Cleaning up project
