+++ assets/app/app.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { NewAuthService } from "./auth/new-auth.service";
+import { Component } from '@angular/core';
+import { AuthService } from "./auth/auth.service";
 
 @Component({
     selector: 'app-root',
@@ -7,7 +7,7 @@ import { NewAuthService } from "./auth/new-auth.service";
 })
 export class AppComponent {
 
-    constructor(public auth: NewAuthService) {
+    constructor(public auth: AuthService) {
         auth.handleAuthentication();
     }
