+++ assets/app/app.module.ts
@@ -3,38 +3,20 @@ import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/cor
 import { HttpModule } from "@angular/http";
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-// Material design
-import {
-    MatSidenavModule,
-    MatCardModule,
-    MatExpansionModule,
-    MatIconModule,
-    MatListModule,
-    MatToolbarModule,
-    MatButtonToggleModule
-} from '@angular/material';
-
 import { AppComponent } from './app.component';
-import { SidenavComponent } from './sidenav/sidenav.component';
-import { AuthenticationComponent } from "./auth/authentication.component";
-import { HeaderComponent } from "./header/header.component";
 import { appRouting } from "./app.routing";
 import { ErrorComponent } from "./errors/error.component";
 import { ErrorService } from "./errors/error.service";
-import { PostModule } from "./posts/post.module";
 
 import { MDBBootstrapModule } from 'angular-bootstrap-md';
-import { WebAppComponent } from "./webapp.component";
 import { LandingPageModule } from "./landing-page/landing-page.module";
-import { NewAuthService } from "./auth/new-auth.service";
+import { AuthService } from "./auth/auth.service";
 import { AuthGuardService } from "./auth/auth-guard.service";
 import { WebAppModule } from "./webapp.module";
-import { RouterModule } from "@angular/router";
 
 @NgModule({
     declarations: [
         AppComponent,
-        AuthenticationComponent,
         ErrorComponent,
     ],
     imports: [
@@ -47,7 +29,7 @@ import { RouterModule } from "@angular/router";
         MDBBootstrapModule.forRoot(),
     ],
     schemas: [ CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA ],
-    providers: [AuthGuardService, NewAuthService, ErrorService],
+    providers: [AuthGuardService, AuthService, ErrorService],
     bootstrap: [AppComponent]
 })
 export class AppModule { }
