+++ assets/app/app.routing.ts
@@ -4,10 +4,10 @@ import { LandingPageComponent } from "./landing-page/landing-page.component";
 import { WebAppComponent } from "./webapp.component";
 
 const APP_ROUTES: Routes = [
-    { path: '', redirectTo: 'home', pathMatch: 'full'},
+    { path: '', redirectTo: 'landing-page', pathMatch: 'full'},
     { path: 'landing-page', component: LandingPageComponent },
     { path: 'home', component: WebAppComponent, canActivate: [AuthGuard], loadChildren: './webapp.module#WebAppModule'},
     { path: '**', redirectTo: 'home', pathMatch: 'full'},
 ];
 
-export const appRouting = RouterModule.forRoot(APP_ROUTES, { enableTracing: true });
+export const appRouting = RouterModule.forRoot(APP_ROUTES);
