--- assets/app/auth/authentication.component.html
@@ -1,12 +1,0 @@
-<header class="row-spacing">
-    <nav class="col-md-8 col-md-offset-2">
-        <ul class="nav nav-tabs">
-            <li routerLinkActive="active"><a [routerLink]="['sign-up']">Sign Up</a></li>
-            <li routerLinkActive="active" *ngIf="!isLoggedIn()"><a [routerLink]="['sign-in']">Sign In</a></li>
-            <li routerLinkActive="active" *ngIf="isLoggedIn()"><a [routerLink]="['logout']">Logout</a></li>
-        </ul>
-    </nav>
-</header>
-<div class="row spacing">
-    <router-outlet></router-outlet>
-</div>
