--- assets/app/auth/authentication.component.ts
@@ -1,17 +1,0 @@
-import { Component } from "@angular/core";
-import { AuthService } from "./auth.service";
-
-@Component({
-    selector: 'app-authentication',
-    templateUrl: './authentication.component.html'
-})
-
-export class AuthenticationComponent {
-
-    constructor(private authService: AuthService) {}
-
-    isLoggedIn() {
-        return this.authService.isLoggedIn();
-    }
-
-}
