--- assets/app/auth/sign-in.component.html
@@ -1,27 +1,0 @@
-<div class="col-md-8 col-md-offset-2">
-    <form [formGroup]="myForm" (ngSubmit)="onSubmit()">
-        <div class="form-group">
-            <label for="identification">Username or email address</label>
-            <input
-                    type="text"
-                    id="identification"
-                    class="form-control"
-                    formControlName="identification"
-            >
-        </div>
-        <div class="form-group">
-            <label for="password">Password</label>
-            <input
-                    type="password"
-                    id="password"
-                    class="form-control"
-                    formControlName="password"
-            >
-        </div>
-        <button
-                class="btn btn-primary"
-                type="submit"
-                [disabled]="!myForm.valid"
-        >Submit</button>
-    </form>
-</div>
