--- assets/app/auth/sign-up.component.html
@@ -1,54 +1,0 @@
-<div class="col-md-8 col-md-offset-2">
-    <form [formGroup]="myForm" (ngSubmit)="onSubmit()">
-        <div class="form-group">
-            <label for="username">Username</label>
-            <input
-                    type="text"
-                    id="username"
-                    class="form-control"
-                    formControlName="username"
-            >
-        </div>
-        <div class="form-group">
-            <label for="firstName">First Name</label>
-            <input
-                    type="text"
-                    id="firstName"
-                    class="form-control"
-                    formControlName="firstName"
-            >
-        </div>
-        <div class="form-group">
-            <label for="lastName">Last Name</label>
-            <input
-                    type="text"
-                    id="lastName"
-                    class="form-control"
-                    formControlName="lastName"
-            >
-        </div>
-        <div class="form-group">
-            <label for="email">Email Address</label>
-            <input
-                    type="email"
-                    id="email"
-                    class="form-control"
-                    formControlName="email"
-            >
-        </div>
-        <div class="form-group">
-            <label for="password">Password</label>
-            <input
-                    type="password"
-                    id="password"
-                    class="form-control"
-                    formControlName="password"
-            >
-        </div>
-        <button
-                class="btn btn-primary"
-                type="submit"
-                [disabled]="!myForm.valid"
-        >Submit</button>
-    </form>
-</div>
