--- assets/app/auth/sign-up.component.ts
@@ -1,44 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-import { FormControl, FormGroup, Validators } from "@angular/forms";
-import { AuthService } from "./auth.service";
-import { User } from "./user.model";
-
-@Component({
-    selector: 'app-sign-up',
-    templateUrl: './sign-up.component.html'
-})
-
-export class SignUpComponent implements OnInit{
-    myForm: FormGroup;
-
-    constructor(private authService: AuthService) {}
-
-    onSubmit() {
-        const user = new User(
-            this.myForm.value.username,
-            this.myForm.value.email,
-            this.myForm.value.password,
-            this.myForm.value.firstName,
-            this.myForm.value.lastName
-        );
-        this.authService.signUp(user)
-            .subscribe(
-                data => console.log(data),
-                error => console.error(error)
-            );
-        this.myForm.reset();
-    }
-
-    ngOnInit() {
-        this.myForm = new FormGroup({
-            username: new FormControl(null, Validators.required),
-            firstName: new FormControl(null, Validators.required),
-            lastName: new FormControl(null, Validators.required),
-            email: new FormControl(null, [
-                Validators.required,
-                Validators.pattern("[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$")
-            ]),
-            password: new FormControl(null, Validators.required)
-        });
-    }
-}
