--- assets/app/form-extension/validate-equal.ts
@@ -1,11 +1,0 @@
-import { AbstractControl } from "@angular/forms";
-
-export class equalValidation {
-    static matchPassword(AC: AbstractControl){
-        let password=AC.get('password').value;
-        let repeatPassword=AC.get('repeat-password').value;
-        if(password!=repeatPassword){
-            console.log(false);
-        }
-    }
-}
