--- assets/app/google-auth/google-sign-in.component.html
@@ -1,9 +1,0 @@
-<google-signin
-        [clientId]="myClientId"
-        [width]="400"
-        [theme]="dark"
-        [scope]="myScope"
-        [longTitle]="true"
-        (googleSignInSuccess)="onGoogleSignInSuccess($event)">
-</google-signin>
-<button (click)="onSignOut()"></button>
