--- assets/app/google-auth/googleauth.model.ts
@@ -1,7 +1,0 @@
-export class GoogleAuthModel {
-    token: string;
-
-    constructor(token: string) {
-        this.token = token;
-    }
-}
