+++ assets/app/header/user-options/user-options-dropdown.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from "@angular/core";
-import { NewAuthService } from "../../auth/new-auth.service";
+import { AuthService } from "../../auth/auth.service";
 
 @Component({
     selector: 'app-user-options-dropdown',
@@ -10,7 +10,7 @@ import { NewAuthService } from "../../auth/new-auth.service";
 export class UserOptionsDropdownComponent {
     @Input() profile: any;
 
-    constructor(private authService: NewAuthService) {}
+    constructor(private authService: AuthService) {}
 
     hasImage(): boolean {
         if (this.profile) {
