+++ assets/app/landing-page/landing-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { NewAuthService } from "../auth/new-auth.service";
+import { AuthService } from "../auth/auth.service";
 import { Router } from "@angular/router";
 const image = require('../../images/350214-mountain-view.jpg');
 
@@ -11,10 +11,10 @@ const image = require('../../images/350214-mountain-view.jpg');
 
 export class LandingPageComponent implements OnInit{
 
-    constructor(private newAuthService: NewAuthService, private router: Router) {}
+    constructor(private authService: AuthService, private router: Router) {}
 
     ngOnInit(): void {
-        if (!this.newAuthService.isAuthenticated()) this.newAuthService.login();
+        if (!this.authService.isAuthenticated()) this.authService.login();
         else this.router.navigate(['home'])
     }
